splitting things out
